generate hash task - only when file exists

+++ BuildBySignature/GenerateHashTask.cs
@@ -22,33 +22,36 @@ namespace BuildBySignature
 
 		public override bool Execute()
 		{
-			var hash = Hasher.Hash(TargetPath, Log).ToString("X8");
-			string currentHash = null;
-			string currentStamp = null;
-			if (File.Exists(BbsPath))
+			if (File.Exists(TargetPath))
 			{
-				// compatibility - back to normal files due to performance reasons
-				if (File.GetAttributes(BbsPath) != FileAttributes.Normal)
+				var hash = Hasher.Hash(TargetPath, Log).ToString("X8");
+				string currentHash = null;
+				string currentStamp = null;
+				if (File.Exists(BbsPath))
 				{
-					File.SetAttributes(BbsPath, FileAttributes.Normal);
+					// compatibility - back to normal files due to performance reasons
+					if (File.GetAttributes(BbsPath) != FileAttributes.Normal)
+					{
+						File.SetAttributes(BbsPath, FileAttributes.Normal);
+					}
+					var current = File.ReadAllLines(BbsPath);
+					currentHash = current[0];
+					currentStamp = current[1];
 				}
-				var current = File.ReadAllLines(BbsPath);
-				currentHash = current[0];
-				currentStamp = current[1];
-			}
-			var originalDllLastWriteTime = File.GetLastWriteTimeUtc(TargetPath);
-			if (string.Equals(hash, currentHash, StringComparison.Ordinal) && RevertTargetStamp && !string.IsNullOrEmpty(currentStamp))
-			{
-				File.SetLastWriteTimeUtc(TargetPath, DateTime.ParseExact(currentStamp, "o", CultureInfo.InvariantCulture).ToUniversalTime());
-				Log.LogMessage(MessageImportance.High, " * Bbs GenerateHashTask: Target LastWriteTime Reverted! {0} (#{1} {2})", Path.GetFileName(TargetPath), currentHash, currentStamp);
-			}
-			else
-			{
-				File.WriteAllText(BbsPath, hash + Environment.NewLine + File.GetLastWriteTimeUtc(TargetPath).ToString("o", CultureInfo.InvariantCulture));
-			}
+				var originalDllLastWriteTime = File.GetLastWriteTimeUtc(TargetPath);
+				if (string.Equals(hash, currentHash, StringComparison.Ordinal) && RevertTargetStamp && !string.IsNullOrEmpty(currentStamp))
+				{
+					File.SetLastWriteTimeUtc(TargetPath, DateTime.ParseExact(currentStamp, "o", CultureInfo.InvariantCulture).ToUniversalTime());
+					Log.LogMessage(MessageImportance.High, " * Bbs GenerateHashTask: Target LastWriteTime Reverted! {0} (#{1} {2})", Path.GetFileName(TargetPath), currentHash, currentStamp);
+				}
+				else
+				{
+					File.WriteAllText(BbsPath, hash + Environment.NewLine + File.GetLastWriteTimeUtc(TargetPath).ToString("o", CultureInfo.InvariantCulture));
+				}
 
-			// bbs timestamps should be related to dll timestamps so that it would be possible to detect dll content change by bbs modification date
-			File.SetLastWriteTimeUtc(BbsPath, originalDllLastWriteTime);
+				// bbs timestamps should be related to dll timestamps so that it would be possible to detect dll content change by bbs modification date
+				File.SetLastWriteTimeUtc(BbsPath, originalDllLastWriteTime);
+			}
 			return true;
 		}
